+++ Classes/Domain/Model/Event/EventDate.php
@@ -21,6 +21,11 @@ class EventDate extends Event implements EventDateInterface
 
     public function __construct()
     {
+        $this->initializeObject();
+    }
+
+    public function initializeObject(): void
+    {
         $this->initializeEventDate();
     }