+++ Classes/Domain/Model/Event/EventTopic.php
@@ -14,6 +14,11 @@ class EventTopic extends Event implements EventTopicInterface
 
     public function __construct()
     {
+        $this->initializeObject();
+    }
+
+    public function initializeObject(): void
+    {
         $this->initializeEventTopic();
     }